rename: slack configuration
