+++ example/src/config/slack.ts
